requirements.txt
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
cryptography==36.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
pluggy==1.0.0
py==1.10.0
pycparser==2.21
pyfirecrest==0.0.12
PyJWT==2.1.0
pymongo==4.0.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
toml==0.10.2
tornado==6.1
urllib3==1.26.8
Werkzeug==2.0.2
